backend_repo/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
